+++ stream-approvals/approvals-core/src/main/java/com/backbase/stream/mapper/ApprovalMapper.java
@@ -1,6 +1,5 @@
 package com.backbase.stream.mapper;
 
-import com.backbase.dbs.approval.api.service.v2.model.PostApprovalTypeRequest;
 import com.backbase.dbs.approval.api.service.v2.model.PostScopedApprovalTypeRequest;
 import com.backbase.dbs.approval.api.service.v2.model.PresentationPostBulkApprovalTypeAssignmentRequest;
 import com.backbase.stream.approval.model.ApprovalType;
@@ -11,8 +10,7 @@ import org.mapstruct.Mapping;
 @Mapper
 public interface ApprovalMapper {
 
-    PostApprovalTypeRequest mapApprovalType(ApprovalType approvalType);
-
+    @Mapping(source = "serviceAgreementId", target = "creatorServiceAgreementId")
     PostScopedApprovalTypeRequest mapScopedApprovalType(ApprovalType approvalType);
 
     @Mapping(source = "approvalTypeAssignments", target = "approvalTypeAssignments")