+++ stream-approvals/approvals-core/src/main/java/com/backbase/stream/service/ApprovalsIntegrationService.java
@@ -5,7 +5,6 @@ import com.backbase.dbs.approval.api.service.v2.ApprovalTypesApi;
 import com.backbase.dbs.approval.api.service.v2.PoliciesApi;
 import com.backbase.dbs.approval.api.service.v2.PolicyAssignmentsApi;
 import com.backbase.dbs.approval.api.service.v2.model.PostApprovalTypeResponse;
-import com.backbase.dbs.approval.api.service.v2.model.PostPolicyResponse;
 import com.backbase.dbs.approval.api.service.v2.model.PostPolicyServiceApiResponse;
 import com.backbase.stream.approval.model.ApprovalType;
 import com.backbase.stream.approval.model.Policy;
@@ -13,6 +12,7 @@ import com.backbase.stream.approval.model.PolicyAssignment;
 import com.backbase.stream.exceptions.ApprovalTypeException;
 import com.backbase.stream.exceptions.PolicyAssignmentException;
 import com.backbase.stream.exceptions.PolicyException;
+import com.backbase.stream.legalentity.model.ServiceAgreement;
 import com.backbase.stream.mapper.ApprovalMapper;
 import com.backbase.stream.mapper.PolicyMapper;
 import java.util.Objects;
@@ -22,7 +22,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.mapstruct.factory.Mappers;
 import org.springframework.http.HttpRequest;
 import org.springframework.util.CollectionUtils;
-import org.springframework.util.ObjectUtils;
 import org.springframework.web.reactive.function.client.WebClientResponseException;
 import reactor.core.publisher.Mono;
 
@@ -37,25 +36,27 @@ public class ApprovalsIntegrationService {
     private final ApprovalTypeAssignmentsApi approvalTypeAssignmentsApi;
     private final PoliciesApi policiesApi;
     private final PolicyAssignmentsApi policyAssignmentsApi;
-
-    private static final String CREATE_APPROVAL_POLICY_LOG_MESSAGE = "Created approval policy: '{}' with identifier: [{}].";
+    private final AccessGroupService accessGroupService;
 
     public Mono<ApprovalType> createApprovalType(ApprovalType approvalType) {
-        Mono<PostApprovalTypeResponse> apiCallMono;
-
-        if (ObjectUtils.isEmpty(approvalType.getScope())) {
-            apiCallMono = approvalTypesApi.postApprovalType(approvalMapper.mapApprovalType(approvalType));
-        } else {
-            apiCallMono = approvalTypesApi.postScopedApprovalType(approvalMapper.mapScopedApprovalType(approvalType));
-        }
-
-        return apiCallMono
-            .map(PostApprovalTypeResponse::getApprovalType)
-            .map(at -> {
-                String id = at.getId();
-                log.info("Created approval type: {} with identifier: [{}].", at.getName(), id);
-                approvalType.setInternalId(id);
-                return approvalType;
+        return Mono.just(approvalType)
+            .flatMap(inputApprovalType -> {
+                if (inputApprovalType.getServiceAgreementId() == null) {
+                    return approvalTypesApi.postScopedApprovalType(
+                            approvalMapper.mapScopedApprovalType(inputApprovalType))
+                        .map(createdApprovalType -> mapCreatedApprovalType(inputApprovalType, createdApprovalType));
+                } else {
+                    return accessGroupService.getServiceAgreementByExternalId(inputApprovalType.getServiceAgreementId())
+                        .map(ServiceAgreement::getInternalId)
+                        .defaultIfEmpty(inputApprovalType.getServiceAgreementId())
+                        .flatMap(internalId -> {
+                            inputApprovalType.setServiceAgreementId(internalId);
+                            return approvalTypesApi.postScopedApprovalType(
+                                    approvalMapper.mapScopedApprovalType(inputApprovalType))
+                                .map(createdApprovalType -> mapCreatedApprovalType(inputApprovalType,
+                                    createdApprovalType));
+                        });
+                }
             })
             .doOnError(WebClientResponseException.class, this::handleWebClientResponseException)
             .onErrorResume(WebClientResponseException.class, exception ->
@@ -63,36 +64,47 @@ public class ApprovalsIntegrationService {
             .onErrorStop();
     }
 
+    private ApprovalType mapCreatedApprovalType(ApprovalType inputApprovalType,
+        PostApprovalTypeResponse createdApprovalTypeResponse) {
+        String id = createdApprovalTypeResponse.getApprovalType().getId();
+        String name = createdApprovalTypeResponse.getApprovalType().getName();
+        log.info("Created approval type: {} with identifier: [{}].", name, id);
+        inputApprovalType.setInternalId(id);
+        return inputApprovalType;
+    }
+
     public Mono<Policy> createPolicy(Policy policy) {
-        Mono<Policy> createdPolicyMono;
-
-        if (ObjectUtils.isEmpty(policy.getScope())) {
-            createdPolicyMono = policiesApi.postPolicy(policyMapper.mapPolicy(policy))
-                .map(PostPolicyResponse::getPolicy)
-                .map(at -> {
-                    String id = at.getId();
-                    log.info(CREATE_APPROVAL_POLICY_LOG_MESSAGE, at.getName(), id);
-                    policy.setInternalId(id);
-                    return policy;
-                });
-        } else {
-            createdPolicyMono = policiesApi.postScopedPolicy(policyMapper.mapScopedPolicy(policy))
-                .map(PostPolicyServiceApiResponse::getPolicy)
-                .map(at -> {
-                    String id = at.getId();
-                    log.info(CREATE_APPROVAL_POLICY_LOG_MESSAGE, at.getName(), id);
-                    policy.setInternalId(id);
-                    return policy;
-                });
-        }
-
-        return createdPolicyMono
+        return Mono.just(policy)
+            .flatMap(inputPolicy -> {
+                if (inputPolicy.getServiceAgreementId() == null) {
+                    // Directly proceed with system policy creation
+                    return policiesApi.postScopedPolicy(policyMapper.mapScopedPolicy(inputPolicy))
+                        .map(createdPolicyResponse -> mapCreatedPolicy(inputPolicy, createdPolicyResponse));
+                } else {
+                    // Handle local policy creation
+                    return accessGroupService.getServiceAgreementByExternalId(inputPolicy.getServiceAgreementId())
+                        .map(ServiceAgreement::getInternalId)
+                        .defaultIfEmpty(inputPolicy.getServiceAgreementId())
+                        .flatMap(internalId -> {
+                            inputPolicy.setServiceAgreementId(internalId);
+                            return policiesApi.postScopedPolicy(policyMapper.mapScopedPolicy(inputPolicy))
+                                .map(createdPolicyResponse -> mapCreatedPolicy(inputPolicy, createdPolicyResponse));
+                        });
+                }
+            })
             .doOnError(WebClientResponseException.class, this::handleWebClientResponseException)
             .onErrorResume(WebClientResponseException.class, exception ->
                 Mono.error(new PolicyException(policy, "Failed to create Policy", exception)))
             .onErrorStop();
     }
 
+    private Policy mapCreatedPolicy(Policy inputPolicy, PostPolicyServiceApiResponse createdPolicyResponse) {
+        String id = createdPolicyResponse.getPolicy().getId();
+        String name = createdPolicyResponse.getPolicy().getName();
+        log.info("Created approval policy: '{}' with identifier: [{}].", name, id);
+        inputPolicy.setInternalId(id);
+        return inputPolicy;
+    }
 
     public Mono<PolicyAssignment> assignPolicies(PolicyAssignment policyAssignment) {
         return policyAssignmentsApi